Reorganize some CA manager code for correctness/readability
